move all (inc RootChecked) into views/impls.rs, add Rc + Arc
